requirements.txt
pyqt
numpy
pandas
scipy
matplotlib
